+++ packages/variance/src/types/config.ts
@@ -9,7 +9,7 @@ import {
   SelectorProps,
   ThemeProps,
 } from './props';
-import { Key } from './utils';
+import { AllUnionKeys, Key, KeyFromUnion } from './utils';
 
 export type LiteralScale = Record<string | number, string | number>;
 
@@ -38,7 +38,7 @@ export interface AbstractParser {
   config: Record<string, AbstractPropTransformer>;
 }
 
-type PropertyValues<
+export type PropertyValues<
   Property extends keyof PropertyTypes,
   All extends boolean = false
 > = Exclude<PropertyTypes[Property], All extends true ? never : object | any[]>;
@@ -69,6 +69,39 @@ export interface PropTransformer<P extends string, C extends Prop>
 export type TransformerMap<Config extends Record<string, Prop>> = {
   [P in Key<keyof Config>]: PropTransformer<Key<P>, Config[P]>;
 };
+export interface Parser<
+  Config extends Record<string, AbstractPropTransformer>
+> {
+  (props: ParserProps<Config>): CSSObject;
+  propNames: (keyof Config)[];
+  config: Config;
+}
+
+export type Compose<Args extends AbstractParser[]> = {
+  [K in AllUnionKeys<Args[number]['config']>]: KeyFromUnion<
+    Args[number]['config'],
+    K
+  >;
+};
+export interface Variant<P extends AbstractParser> {
+  <
+    Keys extends keyof Props,
+    Base extends AbstractProps,
+    Props extends Record<Keys, AbstractProps>,
+    PropKey extends Readonly<string> = 'variant'
+  >(options: {
+    prop?: PropKey;
+    defaultVariant?: keyof Props;
+    base?: SelectorProps<Base, SystemProps<P>>;
+    variants: SelectorMap<Props, SystemProps<P>>;
+  }): (props: VariantProps<PropKey, Keys | false> & ThemeProps) => CSSObject;
+}
+
+export interface CSS<P extends AbstractParser> {
+  <Props extends AbstractProps>(config: SelectorProps<Props, SystemProps<P>>): (
+    props: ThemeProps
+  ) => CSSObject;
+}
 
 export type ParserProps<
   Config extends Record<string, AbstractPropTransformer>
@@ -80,33 +113,11 @@ export type ParserProps<
   }
 >;
 
-export interface Parser<
-  Config extends Record<string, AbstractPropTransformer>
-> {
-  (props: ParserProps<Config>): CSSObject;
-  propNames: (keyof Config)[];
-  config: Config;
-}
-
 export type SystemProps<P extends AbstractParser> = Omit<
   Parameters<P>[0],
   'theme'
 >;
 
-export interface Variant<P extends AbstractParser> {
-  <
-    Keys extends keyof Props,
-    Props extends Record<string, AbstractProps>,
-    PropKey extends Readonly<string> = 'variant'
-  >(options: {
-    prop?: PropKey;
-    defaultVariant?: Keys;
-    variants: SelectorMap<Props, SystemProps<P>>;
-  }): (props: ThemeProps<Partial<Record<PropKey, Keys>>>) => CSSObject;
-}
-
-export interface CSS<P extends AbstractParser> {
-  <Props extends AbstractProps>(config: SelectorProps<Props, SystemProps<P>>): (
-    props: ThemeProps
-  ) => CSSObject;
-}
+type VariantProps<T extends string, V> = {
+  [Key in T]?: V;
+};